[35] - Test unitario del twitch provider añadido
