+++ tests/Unit/Services/TwitchProviderTest.php
@@ -1,0 +1,36 @@
+<?php
+
+namespace Services;
+
+use App\Services\ApiClient;
+use App\Services\TwitchProvider;
+use Mockery;
+use Tests\TestCase;
+
+class TwitchProviderTest extends TestCase
+{
+    /**
+     * @test
+     */
+    public function getTokenTwitchTest()
+    {
+        $mockery = new Mockery();
+        $apiClient = $mockery->mock(ApiClient::class);
+        $tokenExpected = json_encode([
+            'access_token' => 'u308tesk7yzmi8fe7el28e46dad3a5',
+            'expires_in' => 5175216,
+            'token_type' => 'bearer',
+        ]);
+
+        $apiClient
+            ->expects('getToken')
+            ->with('https://id.twitch.tv/oauth2/token')
+            ->once()
+            ->andReturn($tokenExpected);
+
+        $twitchProvider = new TwitchProvider($apiClient);
+        $tokenTwitchResult = $twitchProvider->getTokenTwitch();
+
+        $this->assertEquals($tokenTwitchResult, 'u308tesk7yzmi8fe7el28e46dad3a5');
+    }
+}
